fix: Split websocket endpoint logic into a dedicated handler/service
