+++ api/main.py
@@ -405,87 +405,9 @@ if settings.ENABLE_WEBSOCKET:
             logger.error("WebSocket error", job_id=job_id, error=str(e))
             await websocket.close(code=1011)
 
-    @app.websocket("/ws/cases/{case_id}/timeline")
-    async def websocket_case_timeline_endpoint(
-        websocket: WebSocket,
-        case_id: int,
-        token: str = Query(None),
-    ):
-        """
-        WebSocket endpoint for real-time case timeline updates.
+    from api.websockets.case_timeline import register_case_timeline_endpoint
 
-        Requires authentication via token query parameter or Sec-WebSocket-Protocol
-        header using the subprotocol `access_token`.
-        """
-        from services.timeline_realtime import timeline_realtime_bus
-
-        auth_token = token
-        requested_protocols = []
-
-        if "sec-websocket-protocol" in websocket.headers:
-            header_val = websocket.headers["sec-websocket-protocol"]
-            requested_protocols = [p.strip() for p in header_val.split(",")]
-            if "access_token" in requested_protocols:
-                idx = requested_protocols.index("access_token")
-                if idx + 1 < len(requested_protocols):
-                    auth_token = requested_protocols[idx + 1]
-
-        if not auth_token:
-            await websocket.close(code=4001, reason="Authentication required")
-            return
-
-        try:
-            from api.auth import verify_token
-            payload = verify_token(auth_token)
-            user_id = payload.get("sub")
-            if not user_id:
-                await websocket.close(code=4003, reason="Invalid token")
-                return
-        except (TokenExpiredError, InvalidTokenError, AuthError):
-            await websocket.close(code=4001, reason="Invalid or expired token")
-            return
-
-        subprotocol = "access_token" if "access_token" in requested_protocols else None
-        await websocket.accept(subprotocol=subprotocol)
-
-        # Security: enforce that authenticated user owns the case
-        # (best-effort: if ownership can't be validated in this layer, we still keep
-        # connection alive but will stop pushing if case is missing)
-        try:
-            from sqlalchemy.orm import Session
-            from database import SessionLocal
-            from db.models import Case as CaseModel
-
-            db: Session = SessionLocal()
-            try:
-                case = db.query(CaseModel).filter(CaseModel.id == case_id).first()
-                if not case or str(case.user_id) != str(user_id):
-                    await websocket.close(code=403, reason="Forbidden: case ownership required")
-                    return
-            finally:
-                db.close()
-        except Exception:
-            await websocket.close(code=1011, reason="Server error")
-            return
-
-        await websocket.send_json({
-            "type": "subscribed",
-            "case_id": case_id,
-        })
-
-        queue = await timeline_realtime_bus.subscribe(case_id)
-        try:
-            while True:
-                raw = await queue.get()
-                # bus publishes json string
-                # send_json expects python object, so deserialize
-                import json
-                payload_obj = json.loads(raw)
-                await websocket.send_json(payload_obj)
-        except Exception:
-            pass
-        finally:
-            await timeline_realtime_bus.unsubscribe(case_id, queue)
+    register_case_timeline_endpoint(app)
 
 
 if __name__ == "__main__":
